+++ corva/app/stream.py
@@ -3,7 +3,7 @@ from typing import Optional, List
 
 from corva.app.base import BaseApp
 from corva.models.stream import StreamContext, Record, StreamEventData
-from corva.event.event import Event
+from corva.event import Event
 from corva.event.loader.stream import StreamLoader
 from corva.state.redis_adapter import RedisAdapter
 from corva.state.redis_state import RedisState
