+++ corva/event/loader/scheduled.py
@@ -4,7 +4,7 @@ from itertools import chain
 from typing import List
 
 from corva.models.scheduled import ScheduledEventData
-from corva.event.event import Event
+from corva.event import Event
 from corva.event.loader.base import BaseLoader
