+++ corva/event/loader/stream.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import List
 
-from corva.event.event import Event
+from corva.event import Event
 from corva.models.stream import StreamEventData
 from corva.event.loader.base import BaseLoader
