+++ corva/utils.py
@@ -1,4 +1,4 @@
-from corva.event.event import Event
+from corva.event import Event
 
 
 class GetStateKey:
