+++ tests/test_utils.py
@@ -1,5 +1,5 @@
 from corva.models.base import BaseEventData
-from corva.event.event import Event
+from corva.event import Event
 from corva.utils import GetStateKey
 
 PROVIDER = 'provider'
